+++ src/core/sentinel/index.ts
@@ -1,2 +1,0 @@
-export * from "./SentinelResult";
-export * from "./NodeSentinel";