+++ src/implementation/diff/viewer/index.ts
@@ -1,3 +1,6 @@
+export * from "./AbstractDiffViewer";
+export * from "./ComparingBasedDiffViewer";
 export * from "./TreeHierarchyDiffViewer";
 export * from "./NodeMutationDiffViewer";
 export * from "./SubtreeShapeDiffViewer";
+export * from "./CompositeDiffViewer";